fix: refactor code
